added embedded output

+++ init.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 GRADLE_VERSION=7.4.1
-curl -O https://downloads.gradle-dn.com/distributions/gradle-${GRADLE_VERSION}-bin.zip
-unzip  gradle-${GRADLE_VERSION}-bin.zip && rm  gradle-${GRADLE_VERSION}-bin.zip
-GRADLE_HOME=/gradle-${GRADLE_VERSION}/bin
-ENV PATH=${GRADLE_HOME}:$PATH
+if [ ! -e gradle-${GRADLE_VERSION} ];then
+    curl -O https://downloads.gradle-dn.com/distributions/gradle-${GRADLE_VERSION}-bin.zip
+    unzip  gradle-${GRADLE_VERSION}-bin.zip && rm  gradle-${GRADLE_VERSION}-bin.zip
+fi
+GRADLE_HOME=/home/compf/Documents/hibernate/HibernateDemo/gradle-${GRADLE_VERSION}/bin
+export PATH=${GRADLE_HOME}:$PATH
+echo $PATH
